stucture reorganize
