+++ src/lib/services/asset-info/index.ts
@@ -1,0 +1,5 @@
+import type { AssetInfo } from "./asset-service";
+
+export type IAssetService = {
+  getAssetInfo(assetId: string): Promise<AssetInfo>
+}
