+++ src/lib/services/balance/index.ts
@@ -1,10 +1,5 @@
-import type { AssetInfo } from "./asset-service";
+import type { INodeClient } from "$lib/services/node-client";
 import type { BalanceType } from "./balance-service";
-
-export interface INodeClient {
-  getAddressBalance(address: string, assetId: string): Promise<BigInt>;
-  getAssetDetails(assetId: string): Promise<AssetInfo>
-}
 
 export type IBalanceService = {
   nodeClient: INodeClient;
@@ -12,6 +7,4 @@ export type IBalanceService = {
   subscribe(balanceType: BalanceType, assetId: string): ReadableStreamDefaultReader<bigint>;
 };
 
-export type IAssetService = {
-  getAssetInfo(assetId: string): Promise<AssetInfo>
-}
+
