+++ src/lib/services/node-client/index.ts
@@ -1,0 +1,6 @@
+import type { AssetInfo } from "../asset-info/asset-service";
+
+export interface INodeClient {
+  getAddressBalance(address: string, assetId: string): Promise<BigInt>;
+  getAssetDetails(assetId: string): Promise<AssetInfo>
+}
